refactor: 관리자 페이지 상품 관리 영역, 쿠폰 관리 영역 컴포넌트로 분리
