+++ src/refactoring/components/coupon/index.ts
@@ -1,3 +1,4 @@
 export { CouponList } from './CouponList';
+export { CouponManagement } from './CouponManagement';
 export { CouponSelect } from './CouponSelect';
 export { NewCouponForm } from './NewCouponForm';
