+++ src/refactoring/components/product/index.ts
@@ -3,3 +3,4 @@ export { NewProductForm } from './NewProductForm';
 export { ProductAccordion } from './ProductAccordion';
 export { ProductCard } from './ProductCard';
 export { ProductDiscountEditor } from './ProductDiscountEditor';
+export { ProductManagement } from './ProductManagement';
